refactor(queries): split boards.ts into reads + mutations (#135, #143)
Extract reads (rowToBoard, fetchBoards, fetchBoard, isNotFoundError,
useBoards, useBoard, query keys) into boards.read.ts and mutations
(useBoardPatch, useRename/useSetBoardKind/useSetLabelsVisible/
useSetVoiceMode/useSetStepIds/useSetKidReorderable, useCreateBoard) into
boards.mutations.ts. boards.ts becomes a barrel re-export so the 11
import sites at @/lib/queries/boards don't churn.

While moving useSetStepIds, tighten the docstring to name both reasons
the wrapper exists rather than a plain useBoardPatch consumer:
useBoardPatch.toRowPatch can't see current (so update closures don't
fit), and retry must re-derive against the current cache (not a saved
snapshot) because outbox drains and other tabs can shift step_ids
between mutate and retry. This pins the constraint #143 questioned.

Pure refactor: zero behavior change, zero call-site changes. All 268
tests pass.
